requirements.txt
python-dotenv
psycopg2-binary
torch
transformers
requests
